Former-commit-id: b79a100d7eca2affbd0fb7ea19b94e76109fdc84

+++ cocos2d/CCTexturePVR.m
@@ -261,7 +261,7 @@ typedef struct _PVRTexHeader
 					heightBlocks = 2;
 
 				dataSize = widthBlocks * heightBlocks * ((blockSize  * bpp) / 8);
-				float packetLength = (dataLength-dataOffset);
+				unsigned int packetLength = (dataLength-dataOffset);
 				packetLength = packetLength > dataSize ? dataSize : packetLength;
 
 				mipmaps_[numberOfMipmaps_].address = bytes+dataOffset;
